also expose internal shared functionality for external usage
