+++ package.json
@@ -1,5 +1,4 @@
 {
-  "packageManager": "pnpm@8.10.1",
   "scripts": {
     "dev:docker": "docker-compose -f docker-compose.dev.yaml up redis",
     "dev:packages": "pnpm --parallel run dev",
@@ -17,5 +16,6 @@
     "@types/node": "^20.8.10",
     "npm-run-all": "^4.1.5",
     "shelljs": "^0.8.5"
-  }
+  },
+  "packageManager": "pnpm@8.15.4"
 }
